Completion
The app misbehaves: I left the best result of positioning captions I could set, but the lines are way too narrow. Also, the sharing functionality extracts a black rectangle. I did all I could at the moment to spot the cause, but I failed. I hope I will be able to fix it in the future.